requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
fastapi
uvicorn
pyyaml
flake8
pytest
